Adding optional value for worker hostname discovery
This change is to help mitigate microk8s issue https://github.com/canonical/microk8s/issues/1566.
Essentially any pod not running on an api server node will not have DNS/hostname discovery. Commands like "kubectl logs <pod>" will fail due to the api servers not being able to resolve the hostname of the workers in a cluster not attached to a network with a DNS/L3 resolver.